packages/common/pyproject.toml
[project]
name = "common"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "boto3>=1.39.5",
    "docker>=7.1.0",
    "psycopg[binary]>=3.2.10",
    "openapi-generator-cli>=7.14.0",
    "pydantic-settings>=2.10.1",
]

[dependency-groups]
dev = [
    "boto3-stubs[batch,ecs,secretsmanager,s3,lambda]>=1.40.29",
    "types-docker>=7.1.0.20250907",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/common"]
force-include = { "src/common/py.typed" = "common/py.typed" }
